requirements.txt
elevenlabs>=1.0.0
python-dotenv>=1.0.0
google-genai>=0.5.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
